Moved some classes to packages.
