+++ Test/CreatureCardTest.java
@@ -1,3 +1,4 @@
+import cards.CreatureCard;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
