+++ Test/GUITest.java
@@ -1,3 +1,5 @@
+import cards.Card;
+import cards.CreatureCard;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
