--- src/main/java/CardType.java
@@ -1,2 +1,0 @@
-public class CardType {
-}
