+++ src/main/java/cards/MagicCard.java
